refactor: adding typescript, eslint and prettier
