+++ prettierrc.js
@@ -1,0 +1,7 @@
+module.exports =  {
+    semi:  true,
+    trailingComma:  'all',
+    singleQuote:  true,
+    printWidth:  120,
+    tabWidth:  4,
+  };
