+++ src/components/Square.tsx
@@ -1,5 +1,10 @@
 import React from 'react';
-export default class Square extends React.Component {
+
+interface IProps {
+  value: number,
+  onClick: React.MouseEventHandler<HTMLButtonElement>
+}
+export default class Square extends React.Component<IProps> {
   render() {
     return (
       <button className="square" onClick={this.props.onClick}>
